task(SDK-5045) - Small Icon Fix
