+++ clevertap-pushtemplates/src/main/java/com/clevertap/android/pushtemplates/content/ManualCarouselContentView.kt
@@ -37,7 +37,7 @@ internal class ManualCarouselContentView(
         setCustomBackgroundColour(baseContent.colorData.backgroundColor, R.id.content_view_big)
         setCustomTextColour(baseContent.colorData.titleColor, R.id.title)
         setCustomTextColour(baseContent.colorData.messageColor, R.id.msg)
-        setCustomContentViewSmallIcon(baseContent.iconData.smallIconBitmap, renderer.smallIcon)
+        setCustomContentViewSmallIcon(renderer.smallIconBitmap, renderer.smallIcon)
         setCustomContentViewLargeIcon(baseContent.iconData.largeIcon)
 
         val scaleType = data.carouselData.scaleType
